second change

+++ app/Http/Controllers/admin/ProductController.php
@@ -26,7 +26,9 @@ class ProductController extends Controller
     }
     public function Add(){
         $data['header_title']="Add New Product";
-        $data['getCategory']=CategoryModel::getRecord();
+        $data['getCategory']=CategoryModel::getRecordActive();
+        $data['getBrand']=BrandModel::getRecordActive();
+        $data['getColor']=ColorModel::getRecordActive();
         return view("admin.product.add",$data);
     }
     public function Insert(Request $request){
@@ -36,19 +38,23 @@ class ProductController extends Controller
        if($validator->passes()){
         $product=new ProductModel();
         $title=trim($request->title);
+        $title=trim($request->title);
         $product->title=$title;
-        $product->created_by=Auth::user()->id;
-        $product->category_id=1;
-        $product->sub_category_id=1;
-        $product->brand_id=1;
-        $product->created_by=1;
+        //$product->category_id=$request->category_id;
+        $product->sub_category_id=$request->sub_category_id;
+        $product->brand_id=$request->brand_id;
+        if($request->old_price==null)
+        {
+            $old_price=0;
+        }
         $product->old_price=0;
-        $product->price=0;
-        $product->short_description='';
-        $product->description='';
-        $product->additional_information='';
-        $product->shipping_returns='';
-        $product->sku='';
+        $product->price=$request->price;
+        $product->short_description=trim($request->short_description);
+        $product->description=$request->input('description');
+        $product->additional_information=$request->input('additional_information');
+        $product->shipping_returns=$request->input('shipping_return');
+        $product->status=$request->status;
+        $product->sku=$request->sku;
         $product->save();
 
         
@@ -62,8 +68,49 @@ class ProductController extends Controller
             $product->slug=$new_slug;
             $product->save();
         }
-        session()->flash("success","Product successfully created");
-        return redirect(route("admin.product.edit",$product->id));
+        if(!empty($request->color_id)){
+
+            foreach($request->color_id as $color_id){
+                $color=new ProductColorModel;
+                $color->product_id=$product->id;
+                $color->color_id=$color_id;
+                $color->save();
+            }
+        }
+        $quantity=0;
+        if(!empty($request->size)){
+            foreach($request->size as $size){
+                if(!empty($size['name'])){
+                    $sizeSave=new ProductSizeModel;
+                    $sizeSave->name=$size['name'];
+                    $sizeSave->quantity=$size['quantity'];
+                    $quantity+=$size['quantity'];
+                    $sizeSave->price=!empty($size['price'])?$size['price']:0;
+                    $sizeSave->product_id=$product->id;
+                    $sizeSave->save();
+                }
+            }
+        }
+        $product->quantity=$quantity;
+        $product->save();
+
+        if(!empty($request->file('image'))){
+            foreach($request->file('image') as $value){
+                if($value->isValid()){
+                    $ext=$value->getClientOriginalExtension();
+                    $randomStr=$product->id.Str::random(10);
+                    $filename=strtolower($randomStr).'.'.$ext;
+                    $value->move(public_path('/upload/product/'),$filename);
+                    $imageUpload=new ProductImageModel;
+                    $imageUpload->image_name=$filename;
+                    $imageUpload->image_extension=$ext;
+                    $imageUpload->product_id=$product->id;
+                    $imageUpload->save();
+                }
+            }
+        }
+        session()->flash('success',"Updated successfully");
+        return redirect()->route('admin.product.list');
        }else{
             return back()
             ->withErrors($validator)
@@ -77,17 +124,23 @@ class ProductController extends Controller
             $data['getCategory']=CategoryModel::getRecordActive();
             $data['getBrand']=BrandModel::getRecordActive();
             $data['getColor']=ColorModel::getRecordActive();
-            $data['getSubCategory']=SubCategoriesModel::getRecordSubCategory($product->category_id);
+            $subCategory=SubCategoriesModel::getSingle($product->sub_category_id);
+            $data['getSubCategory']=SubCategoriesModel::getRecordSubCategory($subCategory->category_id);
             $data['header_title']="Edit Product";
             return view("admin.product.edit",$data);
         }else{
             abort(404);
         }
     }
+    public function Detail($id)
+    {
+        $data['getProduct']=ProductModel::getSingle($id);
+        return view("admin.product.detail",$data);
+    }
     public function Update($product_id, Request $request){
         $validator=Validator::make($request->all(),[
             'title'=>'required',
-            'sku'=>'required',
+            'sku'=>'required|unique:product,slug,'.$product_id,
             'price'=>'required|numeric',
             'old_price'=>'required|numeric'
         ]);
@@ -96,8 +149,7 @@ class ProductController extends Controller
             if($validator->passes()){
                 $title=trim($request->title);
                 $product->title=$title;
-                $product->created_by=Auth::user()->id;
-                $product->category_id=$request->category_id;
+                //$product->category_id=$request->category_id;
                 $product->sub_category_id=$request->sub_category_id;
                 $product->brand_id=$request->brand_id;
                 $product->old_price=$request->old_price;
@@ -111,20 +163,22 @@ class ProductController extends Controller
                 $product->save();
                 $slug=Str::slug($title,'-');
                 $checkSlug=ProductModel::checkSlug($slug);
+               
                 if(empty($checkSlug)){
-                        $product->slug=$slug;
-                        $product->save();
+                    $product->slug=$slug;
+                    $product->save();
                 }else{
                     if(ProductModel::getSingle($product_id)->slug==$slug){
                         $product->slug=$slug;
                         $product->save();
                     }
-                   else{
+                   else{ 
                     $new_slug=$slug.'-'.$product->id;
                     $product->slug=$new_slug;
                     $product->save();
                    }
                 }
+                
                 ProductColorModel::DeleteRecord($product->id);
                 if(!empty($request->color_id)){
 
@@ -135,18 +189,23 @@ class ProductController extends Controller
                         $color->save();
                     }
                 }
+                $quantity=0;
                 ProductSizeModel::DeleteRecord($product_id);
                 if(!empty($request->size)){
                     foreach($request->size as $size){
                         if(!empty($size['name'])){
                             $sizeSave=new ProductSizeModel;
                             $sizeSave->name=$size['name'];
+                            $sizeSave->quantity=$size['quantity'];
+                            $quantity+=$size['quantity'];
                             $sizeSave->price=!empty($size['price'])?$size['price']:0;
                             $sizeSave->product_id=$product->id;
                             $sizeSave->save();
                         }
                     }
                 }
+                $product->quantity=$quantity;
+                $product->save();
                 if(!empty($request->file('image'))){
                     foreach($request->file('image') as $value){
                         if($value->isValid()){
@@ -172,13 +231,31 @@ class ProductController extends Controller
             abort(404);
         }
     }
+    public function Remove($id)
+    {
+        $product=ProductModel::getSingle($id);
+        if($product==null){
+            abort(404);
+        }else{
+            $product->delete();
+            session()->flash("success","Category successfully deleted");
+            return redirect()->back();
+        }
+        
+    }
     public function DeleteImage($id){
         $image=ProductImageModel::getSingle($id);
-        if(!empty($image->getLogo())){
-            unlink(public_path('/upload/product/'.$image->image_name));
+        if($image==null)
+        {
+            abort(404);
+        }else{
+            if(!empty($image->getLogo())){
+                unlink(public_path('/upload/product/'.$image->image_name));
+            }
+            $image->delete();
+            return redirect()->back()->with('success','Deleted product image successfully');
         }
-        $image->delete();
-        return redirect()->back()->with('success','Deleted product image successfully');
+        
     }
     public function Sortable(Request $request){
         if(!empty($request->photo_id)){
